web/backend/requirements.txt
aiohttp
aiomysql
aioredis
bcrypt
coloredlogs
minio
msgpack
networkx
pyjwt
tornado
watchgod
